moved global grid code to a separate file to avoid confusion. Also disabled parallel tests for SetEta since it's column-based
